requirements.txt
Django==3.1
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
asgiref==3.2.10
sqlparse==0.2.2
pytz==2020.1
Django-ckeditor==5.9.0
